requirements.txt
msrestazure
azure-mgmt-subscription
azure-mgmt-compute
azure-identity
